refactor: auth pages
