+++ src/components/ProfileHero.jsx
@@ -1,9 +1,6 @@
-import { useEffect, useState } from "react";
-import { useRouter } from "next/navigation";
+import { useState } from "react";
 import { twMerge } from "tailwind-merge";
 import { useDispatch, useSelector } from "react-redux";
-
-import { toast } from "react-toastify";
 
 import ProfileIcon from "@/components/ProfileIcon";
 import Dots from "@/components/Patterns/Dots";
@@ -11,7 +8,6 @@ import Container from "@/components/Container";
 import Tag from "@/components/Tag";
 import Button from "@/components/Button";
 import Modal from "@/components/Modal";
-import Text from "@/components/Text";
 import { inputStyles } from "./Input";
 import {
   getUserProfile,
@@ -19,14 +15,12 @@ import {
 } from "../app/redux/features/profile/profileSlice";
 import useAuth from "@/hooks/useAuth";
 import { selectAuth } from "@/app/redux/features/profile/authSlice";
-import Secure from "@/utils/SecureLs";
+import Form from "./Form";
 
 const EditProfileForm = ({ data, setIsOpen }) => {
-  const [isLoading, setIsloading] = useState(false);
-  const { isAuthenticated, user, token } = useSelector(selectAuth);
+  const { token } = useSelector(selectAuth);
 
   const dispatch = useDispatch();
-  const router = useRouter();
 
   const [profileData, setProfileData] = useState({
     first_name: data?.first_name || "",
@@ -52,13 +46,13 @@ const EditProfileForm = ({ data, setIsOpen }) => {
     Object.entries(profileData).forEach(([key, value]) => {
       formData.append(key, value);
     });
-    await dispatch(updateProfile({ token, formData }));
+    dispatch(updateProfile({ token, formData }));
     dispatch(getUserProfile(token));
     setIsOpen(false);
   };
 
   return (
-    <form
+    <Form
       className="max-w-md mx-auto"
       onSubmit={handleSubmit}
       encType="multipart/form-data"
@@ -112,12 +106,11 @@ const EditProfileForm = ({ data, setIsOpen }) => {
           Update
         </Button>
       </div>
-    </form>
+    </Form>
   );
 };
 
 const ProfileHero = ({ userInfo }) => {
-  const [currentUser, setCurrentUser] = useState(true);
   const [isOpen, setIsOpen] = useState(false);
   const { user } = useAuth();
 
@@ -125,26 +118,25 @@ const ProfileHero = ({ userInfo }) => {
     <div className="bg-queen-blue text-white relative pt-28 pb-20 overflow-hidden">
       <Container size="4xl" className="space-y-4">
         <ProfileIcon photoUrl={userInfo?.imageUrl} className="h-20 w-20" />
-        {currentUser && (
-          <>
-            <Button
-              as="button"
-              onClick={() => setIsOpen(true)}
-              type="button"
-              variant="yellow"
-            >
-              Edit Profile
-            </Button>
+        <>
+          <Button
+            as="button"
+            onClick={() => setIsOpen(true)}
+            type="button"
+            variant="yellow"
+          >
+            Edit Profile
+          </Button>
 
-            <Modal
-              open={isOpen}
-              onClose={() => setIsOpen(false)}
-              heading="Edit profile"
-            >
-              <EditProfileForm data={userInfo} setIsOpen={setIsOpen} />
-            </Modal>
-          </>
-        )}
+          <Modal
+            open={isOpen}
+            onClose={() => setIsOpen(false)}
+            heading="Edit profile"
+          >
+            <EditProfileForm data={userInfo} setIsOpen={setIsOpen} />
+          </Modal>
+        </>
+
         <div className="max-w-96">
           <h1 className="font-heading uppercase text-2xl">
             {`${userInfo?.first_name} ${userInfo?.last_name}`}
